requirements.txt
ai-engine-sdk==0.1.0
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.3.0
anyio==4.4.0
apispec==6.6.1
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
audioread==3.0.1
backcall==0.2.0
backoff==2.2.1
base58==2.1.1
beautifulsoup4==4.12.3
bech32==1.2.0
bleach==6.1.0
blinker==1.8.2
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cosmpy==0.9.2
cycler==0.12.1
cytoolz==0.12.3
dash==2.17.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.7
datasets==2.21.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docopt==0.6.2
ecdsa==0.18.0
eth-hash==0.7.0
eth-keys==0.5.1
eth-typing==4.4.0
eth-utils==2.3.1
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.0
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google_search_results==2.4.2
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==8.0.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
llama-cloud==0.0.11
llama-index==0.10.58
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.58
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.27
llama-index-multi-modal-llms-openai==0.1.8
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.31
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
llvmlite==0.43.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
multitasking==0.0.11
mutagen==1.47.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
newsapi-python==0.2.7
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
openai==0.28.0
openai-whisper==20231117
opencage==2.4.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
peewee==3.17.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.2
plotly==5.22.0
pooch==1.8.2
prompt_toolkit==3.0.47
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
py-bip39-bindings==0.1.11
py-ed25519-zebra-bindings==1.0.1
py-sr25519-bindings==0.2.0
pyarrow==16.1.0
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==1.10.16
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
PyNaCl==1.5.0
pyparsing==3.1.2
pypdf==4.3.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qrcode==7.4.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.4
scalecodec==1.2.11
scikit-learn==1.5.0
scipy==1.13.1
setuptools==71.0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.4.0
SQLAlchemy==2.0.31
stack-data==0.6.3
streamlit==1.36.0
striprtf==0.0.26
substrate-interface==1.7.10
substrateinterface==1.0.0
sympy==1.13.2
tenacity==8.4.2
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.2.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.44.2
twilio==9.2.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uagents==0.12.2
urllib3==2.2.1
uvicorn==0.20.0
vaderSentiment==3.3.2
watchdog==4.0.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.0
Werkzeug==3.0.3
wrapt==1.16.0
xxhash==3.4.1
yarg==0.1.9
yarl==1.9.4
yfinance==0.2.40
yt-dlp==2024.8.6
zipp==3.19.2
